requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==4.1
django-crispy-forms==1.14.0
django-filter==22.1
django-import-export==2.8.0
django-khipu==1.0.0
django-pwa==1.0.10
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
khipu==0.0.5
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.9.1
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tzdata==2022.1
Unipath==1.1
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
